requirements.txt
discord.py
openai
llama-index
PyGithub
python-dotenv
pytest